+++ spec/rails_helper.rb
@@ -34,6 +34,7 @@ ActiveAdmin.application.current_user_method = false
 
 require_relative "support/active_admin_integration_spec_helper"
 require_relative "support/graphql_deny_post_authorization_adapter"
+require_relative "support/graphql_deny_user_authorization_adapter"
 
 RSpec.configure do |config|
   config.use_transactional_fixtures = true