+++ spec/spec_helper.rb
@@ -27,4 +27,13 @@ RSpec.configure do |config|
   config.example_status_persistence_file_path = ".rspec_failures"
 end
 require "polyrun/rspec"
+Polyrun::RSpec.install_sharded_formatter_compat!
 Polyrun::RSpec.install_failure_fragments!
+Polyrun::RSpec.install_worker_ping!
+Polyrun::RSpec.install_example_debug!
+Polyrun::RSpec.install_example_rails_logging!
+Polyrun::RSpec.install_example_timeout!
+Polyrun::RSpec.install_example_prosopite!
+if %w[1 true yes].include?(ENV["POLYRUN_SPEC_QUALITY"]&.to_s&.downcase)
+  Polyrun::RSpec.install_spec_quality!
+end